Add HardwareSensorMtiRos as a lib (librtslamros)
The MTI hardware was setup to read data from a rostopic, for now
hardcoded to "/uav0/imu". main.hpp was adapted to use this version of
IMU.

The goal is having the IMU reading from a ROS topic in online
mode. But since the camera is not yet setup to work online from
rostopic as well, the offline mode of IMU hardware was set to similar
behaviour as online. It should be reverted later.