Change CAMERA_ID to 1 due to a bug into HardwareCameraSensor
When using mOnlineDump with CAMERA_ID 0 it names the dumped images as
image_<cont>.pgm, while with CAMERA_ID 1 the dumps are named
image_<id>_<cont>.pgm. The problem comes in when trying to run in
mOffline mode from these dumps because it only searches for images
which name contains the id. There's a workaround in RT-SLAM's main.hpp
file added by rtslam's commit ffb346e88bfefbe to use a camera with ID
0 but I think it's better change it to 1. I don't think it will hurt.