Add option parser using boost::program_options
Now it's possible to set options using the command line or a
configuration file. If the same option is set in both command line and
config file the cmd line option will be prefered.

For now only the following options are available:
  --help
  --version
  --replay
  --data-path
  --log-file
but others should be added as needed.

NOTE: technically there will be no distinction between program options
and parameters, but parameters will be prefered to be set in a
configuration file since they change less often than program options.