Add tf broadcaster of estimated robot pose
The broadcast needs to be a pointer because if not it wil try to
initialize the node handler before the ros::init, leading to a runtime
crash.

The parent frame is set to "world" since "uav0/origin" didn't work
(don't know why yet). It's not a problem because they are in the same place.

NOTE: "using namespace boost" was removed because it was causing an
ambiguity between our signal function and boost::signal (added from
tf/transform_broadcaster.h)