Add the display QT and GDHE. Compilable but not tested yet.
NOTE: Will only work if the QT classes use Q_SIGNALS, Q_SLOTS, Q_EMIT
instead of signals, slots, emit and so on, if not it will conflict
with boost::signals added by tf. Rtslam and qdisplay in robotpkg/wip
as today are not in conform with it but a patch should come soon.

For more info, check
http://qt-project.org/doc/qt-4.8/signalsandslots.html#using-qt-with-3rd-party-signals-and-slots
and
http://wiki.ros.org/qt_ros/Tutorials/Mixing%20Qt%20and%20Boost%20Signals